code_brats/requirements.txt
tensorboardx
pyyaml
yacs
einops
SimpleITK
monai
mmcv
timm
pandas
numpy
scikit-image
sklearn
matplotlib
torchio
medpy
h5py
